requirements.txt
fastapi==0.110.0 

uvicorn==0.29.0

python-jose[cryptography]==3.3.0

passlib[bcrypt]==1.7.4 

pydantic==2.6.4

PyJWT==2.8.0

python-multipart

pillow==10.2.0

openpyxl==3.1.2

setuptools==69.5.1

starlette~=0.36.3